Support for classical LHS Boolean binary operations support and tests.
